Refactor Duck Pond embed sender to be a util.
https://github.com/python-discord/bot/issues/667
